update document processor and extract the reader
