+++ runarion-python/src/utils/document_processor.py
@@ -1,17 +1,11 @@
-"""
-Document processor utility for handling file uploads and content extraction.
-Supports PDF, TXT, and DOCX files with intelligent chunking for large documents.
-"""
-
 import logging
 import re
 from pathlib import Path
 from typing import List, Literal, Optional, Tuple, TypedDict, overload
 
-import fitz  # PyMuPDF
 import tiktoken
-from docx import Document
 
+from .document_reader import DocumentReader
 from .get_model_max_token import (
     get_safe_model_max_tokens,
 )
@@ -48,7 +42,7 @@ class ProcessedDocument(TypedDict):
     status: Literal["success"]
     raw_text: str
     cleaned_text: str
-    chunks: List[ChunkWithStart]
+    chunks: List[Chunk]
     metadata: ProcessedDocumentMetadata
 
 
@@ -60,18 +54,15 @@ class FailedProcessedDocument(TypedDict):
 
 class DocumentProcessor:
     """
-    Handles document ingestion, extraction, and chunking for the novel pipeline.
+    Handles document extraction and chunking for the novel pipeline.
     """
-
-    SUPPORTED_EXTENSIONS = {".pdf", ".txt", ".docx", ".doc"}
 
     def __init__(
         self,
-        upload_path: str = "/app/uploads",
         provider: str = "openai",
         model: str = "gpt-4o",
         safety_margin: float | int = 0.5,
-        max_chunk_size: int | None = None,
+        max_chunk_size: Optional[int] = None,
         paragraph_priority: bool = True,
         sentence_overlap: int = 0,
         token_overlap: int = 50,
@@ -80,7 +71,6 @@ class DocumentProcessor:
         Initialize the document processor with configuration options.
 
         Args:
-            upload_path: Directory to save uploaded files
             provider: LLM provider for token counting and limits
             model: LLM model for token counting and limits
             safety_margin: Portion of tokens to reserve, int means number of tokens, float means ratio of model max tokens
@@ -89,8 +79,7 @@ class DocumentProcessor:
             sentence_overlap: Number of sentences to overlap between chunks, used when paragraph_priority is True
             token_overlap: Number of tokens to overlap between chunks, used when paragraph_priority is False
         """
-        self._upload_path = Path(upload_path)
-        self._upload_path.mkdir(exist_ok=True)
+        self.document_reader = DocumentReader(paragraph_priority)
 
         self._provider = provider
         self._model = model
@@ -105,7 +94,9 @@ class DocumentProcessor:
                 provider, model, safety_margin=safety_margin
             )
         except Exception as e:
-            logger.warning(e)
+            logger.warning(
+                f"Failed to get conservative max token limit for {provider}/{model}: {e}."
+            )
             conservative_limit = None
         if max_chunk_size is None:
             if conservative_limit is None:
@@ -121,144 +112,6 @@ class DocumentProcessor:
         self.paragraph_priority = paragraph_priority
         self.sentence_overlap = sentence_overlap
         self.token_overlap = token_overlap
-
-    def extract_text_from_file(self, file_path: str) -> str:
-        """
-        Extract text content from various file formats.
-
-        Args:
-            file_path: Path to the document file
-
-        Returns:
-            Extracted text content
-
-        Raises:
-            FileNotFoundError: If file doesn't exist
-            ValueError: If file format is not supported
-            Exception: If extraction fails
-        """
-        path = Path(file_path)
-
-        if not path.exists():
-            raise FileNotFoundError(f"File not found: {path}")
-
-        extension = path.suffix.lower()
-
-        if extension not in self.SUPPORTED_EXTENSIONS:
-            raise ValueError(f"Unsupported file format: {extension}")
-
-        try:
-            if extension == ".pdf":
-                return self._extract_from_pdf(path)
-            elif extension == ".txt":
-                return self._extract_from_txt(path)
-            elif extension in [".docx", ".doc"]:
-                return self._extract_from_docx(path)
-            else:
-                raise ValueError(f"Unsupported file format: {extension}")
-        except Exception as e:
-            logger.error(f"Failed to extract text from {path}: {str(e)}")
-            raise
-
-    def _extract_from_pdf(self, file_path: Path) -> str:
-        """Extract text from PDF file using PyMuPDF."""
-        if self.paragraph_priority:
-            return self._extract_paragraphs_from_pdf(file_path)
-        else:
-            # Original extraction method for backward compatibility
-            texts = []
-            with fitz.open(str(file_path)) as doc:
-                for page in doc:
-                    text = page.get_text()  # type: ignore
-                    if text.strip():
-                        texts.append(text)
-
-            return "\n".join(texts)
-
-    def _extract_paragraphs_from_pdf(self, file_path: Path) -> str:
-        """
-        Extract text from PDF with enhanced paragraph detection using PyMuPDF structural analysis.
-
-        Args:
-            file_path: Path to PDF file
-
-        Returns:
-            Text with properly preserved paragraph boundaries
-        """
-        endings = [".", "!", "?"]
-        endings = endings + [e + "'" for e in endings] + [e + '"' for e in endings]
-
-        paragraphs = []
-        temp_paragraph = ""
-        with fitz.open(str(file_path)) as doc:
-            for page in doc:
-                blocks = page.get_text("blocks")  # type: ignore
-                for i, block in enumerate(blocks):
-                    # skip if image block
-                    if block[6] != 0:
-                        continue
-                    # extract the text from the block
-                    text: str = block[4].strip()
-                    # skip if empty
-                    if not text:
-                        continue
-                    # handle paragraph break at end of page
-                    if i == len(blocks) - 1 and not text.endswith(tuple(endings)):
-                        temp_paragraph = text
-                        continue
-                    paragraphs.append(temp_paragraph + text)
-                    temp_paragraph = ""
-
-        # Join paragraphs with double line breaks to maintain structure
-        return "\n\n".join(paragraphs)
-
-    def _extract_from_txt(self, file_path: Path) -> str:
-        """Extract text from TXT file with encoding detection."""
-        try:
-            with open(file_path, "r", encoding="utf-8") as file:
-                return file.read()
-        except UnicodeDecodeError:
-            # Try with different encoding
-            with open(file_path, "r", encoding="latin-1") as file:
-                return file.read()
-
-    def _extract_from_docx(self, file_path: Path) -> str:
-        """Extract text from DOCX file."""
-        doc = Document(str(file_path))
-
-        paragraphs = []
-        for paragraph in doc.paragraphs:
-            if paragraph.text.strip():
-                paragraphs.append(paragraph.text)
-
-        return "\n\n".join(paragraphs)
-
-    def clean_text(self, text: str) -> str:
-        """
-        Clean and normalize extracted text.
-
-        Args:
-            text: Raw extracted text
-
-        Returns:
-            Cleaned text
-        """
-        # Remove excessive whitespace
-        text = re.sub(r"\s+", " ", text)
-
-        # Remove page numbers and common PDF artifacts
-        text = re.sub(r"\b\d+\b(?=\s*$)", "", text, flags=re.MULTILINE)
-
-        # Remove excessive line breaks
-        text = re.sub(r"\n\s*\n\s*\n+", "\n\n", text)
-
-        # Remove zero-width characters
-        text = re.sub(r"[\u200b\u200c\u200d\ufeff]", "", text)
-
-        # Strip leading/trailing whitespace
-        text = text.strip()
-
-        return text
 
     def chunk_text(
         self,
@@ -801,6 +654,30 @@ class DocumentProcessor:
         sentences = re.split(r"""(?<=[.!?]"\s)|(?<=[.!?]'\s)|(?<=[.!?]\s)""", text)
         return [s.strip() for s in sentences if s.strip()]
 
+    def add_start_to_chunks(
+        self, text: str, chunks: list[Chunk]
+    ) -> List[ChunkWithStart]:
+        """
+        Add start index to each chunk based on its position in the original text.
+
+        Args:
+            text: Original text content
+            chunks: List of chunk dictionaries without start index
+
+        Returns:
+            List of chunks with start index added
+        """
+        chunks_with_start: List[ChunkWithStart] = []
+        for chunk in chunks:
+            chunks_with_start.append(
+                {
+                    **chunk,
+                    "start": text.find(chunk["raw_text"]),
+                }
+            )
+
+        return chunks_with_start
+
     def process_document(
         self, file_path: str
     ) -> ProcessedDocument | FailedProcessedDocument:
@@ -815,23 +692,13 @@ class DocumentProcessor:
         """
         try:
             # Extract text
-            raw_text = self.extract_text_from_file(file_path)
+            raw_text = self.document_reader.extract(file_path)
 
             # Clean text
-            cleaned_text = self.clean_text(raw_text)
+            cleaned_text = self.document_reader.clean(raw_text)
 
             # Create chunks
             chunks = self.chunk_text(cleaned_text)
-
-            # Add start positions for each chunk
-            chunks_with_start: List[ChunkWithStart] = []
-            for chunk in chunks:
-                chunks_with_start.append(
-                    {
-                        **chunk,
-                        "start": cleaned_text.find(chunk["raw_text"]),
-                    }
-                )
 
             # Calculate metadata
             metadata: ProcessedDocumentMetadata = {
@@ -848,7 +715,7 @@ class DocumentProcessor:
                 "status": "success",
                 "raw_text": raw_text,
                 "cleaned_text": cleaned_text,
-                "chunks": chunks_with_start,
+                "chunks": chunks,
                 "metadata": metadata,
             }
 
@@ -857,25 +724,4 @@ class DocumentProcessor:
             return {"status": "failed", "error": str(e), "file_path": file_path}
 
     def validate_file(self, file_path: str) -> Tuple[bool, str]:
-        """
-        Validate file before processing.
-
-        Args:
-            file_path: Path to file
-
-        Returns:
-            Tuple of (is_valid, error_message)
-        """
-        path = Path(file_path)
-
-        if not path.exists():
-            return False, f"File not found: {path}"
-
-        if path.suffix.lower() not in self.SUPPORTED_EXTENSIONS:
-            return False, f"Unsupported file format: {path.suffix}"
-
-        # Check file size (100MB limit)
-        if path.stat().st_size > 100 * 1024 * 1024:
-            return False, "File too large (max 100MB)"
-
-        return True, ""
+        return self.document_reader.validate_file(file_path)
